Move download button templates into their own folder
Split the helpers into different files

bug 814079
