--- apps/mozorg/templates/mozorg/download_button_large.html
@@ -1,1 +1,0 @@
-{% extends "mozorg/download_button_base.html" %}
