+++ apps/mozorg/templates/mozorg/download_buttons/beta_large.html
@@ -1,4 +1,4 @@
-{% extends "/mozorg/download_button_base.html" %}
+{% extends "/mozorg/download_buttons/base.html" %}
 
 {% block class %}download-button-large beta{% endblock %}
