+++ apps/mozorg/templates/mozorg/download_buttons/channel.html
@@ -1,4 +1,4 @@
-{% extends "mozorg/download_button_base.html" %}
+{% extends "mozorg/download_buttons/base.html" %}
 
 {% block class %}download-button-small{% endblock %}
