+++ apps/mozorg/templates/mozorg/download_buttons/dumb.html
