+++ apps/mozorg/templates/mozorg/download_buttons/large.html
@@ -1,0 +1,1 @@
+{% extends "mozorg/download_buttons/base.html" %}
