+++ apps/mozorg/templates/mozorg/download_buttons/mobile_aurora.html
@@ -1,4 +1,4 @@
-{% extends "mozorg/download_button_mobile_base.html" %}
+{% extends "mozorg/download_buttons/mobile_base.html" %}
 
 {% block class %}download-button-mobile-large aurora{% endblock %}
