+++ apps/mozorg/templates/mozorg/download_buttons/upgrade.html
@@ -1,4 +1,4 @@
-{% extends "mozorg/download_button_small.html" %}
+{% extends "mozorg/download_buttons/small.html" %}
 
 {% block title %}Upgrade Firefox{% endblock %}
